PhD_Admission_Portal--saideepika

+++ springapp/src/main/java/com/examly/springapp/Model/UserModel.java
@@ -9,7 +9,7 @@ public class UserModel {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "user_id")
-    private long userId;
+    private int userId;
 
     @Column(name = "email")
     private String email;
@@ -29,7 +29,7 @@ public class UserModel {
     public UserModel() {
     }
 
-    public UserModel(long userId, String email, String password, String username, String mobileNumber, String userRole) {
+    public UserModel(int userId, String email, String password, String username, String mobileNumber, String userRole) {
         this.userId = userId;
         this.email = email;
         this.password = password;
@@ -38,11 +38,11 @@ public class UserModel {
         this.userRole = userRole;
     }
 
-    public long getUserId() {
+    public int getUserId() {
         return this.userId;
     }
 
-    public void setUserId(long userId) {
+    public void setUserId(int userId) {
         this.userId = userId;
     }
